+++ packages/maps/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ldn-viz-tools-maps
 
+## 0.0.3
+
+### Patch Changes
+
+- db4714c: Replace boilerplate README files
+
 ## 0.0.2
 
 ### Patch Changes