+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,39 @@
 # ldn-viz-tools-ui
 
+## 0.3.0
+
+### Minor Changes
+
+- 32b9dd7: allow checkboxes in a checkboxGroup to be initially checked
+- 32b9dd7: rename "style" prop of Button to "variant"
+
+## 0.2.0
+
+### Minor Changes
+
+- f74b4cc: add sidebar component with tab icons on the left
+- d6ffe1e: add custom colored checkbox and radio buttons
+
+## 0.1.0
+
+### Minor Changes
+
+- d5434ea: implement button component
+- 99de4e1: add component for downloading an SVG as an image file
+- 1c76060: prevent interaction with disabled Button component
+- 29a2beb: Add component for downloading data
+- e960b60: add tab selector component
+
+### Patch Changes
+
+- db4714c: Replace boilerplate README files
+
+## 0.0.3
+
+### Patch Changes
+
+- df43eb9: Export Modal and Select from ui package.
+
 ## 0.0.2
 
 ### Patch Changes